requirements.txt
requests
websocket-client
threading
time